refactor: 기존 controller 게임 실행 로직 LadderGame 클래스로 분리
